# feat(composite+viz+bench): Tier-A polish — composite adders, topo eva…