… README.md
… docs/fa.png
… docs/logic.png
… lib/src/main/java/sim/dot/GraphvizExporter.java